Make parsing more forgiving of non-standard bencoded integers.
Since the original bencode standard doesn't allow floats, some software
have unfortunately decided to bencode floats as integers. This is a
stupid idea, because they could've just used a string. But since
.torrent files with this feature already exist, we should try to handle
them gracefully. This commit allows the parsing of floats & scientific
notation, and adds testing for them, casting them to int64 under the
following rules:
-All floats are rounded to the adjacent integer with the lowest absolute
value  (7.5 is rounded to 7, while -7.5 is rounded to -7).
-Floats that exceed the positive or negative bounds of int64
representation are set as -9223372036854775808.